Consolidate reporting tracking
