+++ src/main/java/com/microsoft/bingads/v13/bulk/BulkDownloadOperation.java
@@ -36,29 +36,18 @@ public class BulkDownloadOperation extends BulkOperation<DownloadStatus> {
      */
     public BulkDownloadOperation(String requestId, ServiceClient<IBulkService> serviceClient, int statusPollIntervalInMilliseconds) {
         this(
-                requestId,
-                null,
-                new HttpClientHttpFileService(),
-                Config.DEFAULT_HTTPCLIENT_TIMEOUT_IN_MS,
-                new SimpleZipExtractor(),
-                serviceClient,
-                statusPollIntervalInMilliseconds);
+                requestId, null,
+                new HttpClientHttpFileService(), Config.DEFAULT_HTTPCLIENT_TIMEOUT_IN_MS, new SimpleZipExtractor(),
+                serviceClient, statusPollIntervalInMilliseconds);
     }
 
     BulkDownloadOperation(
-            String requestId,
-            String trackingId,
-            HttpFileService httpFileService,
-            int downloadHttpTimeoutInMilliseconds,
-            ZipExtractor zipExtractor,
-            ServiceClient<IBulkService> serviceClient,
-            int statusPollIntervalInMilliseconds) {
+            String requestId, String trackingId,
+            HttpFileService httpFileService, int downloadHttpTimeoutInMilliseconds, ZipExtractor zipExtractor,
+            ServiceClient<IBulkService> serviceClient, int statusPollIntervalInMilliseconds) {
     	super(
-                requestId,
-                trackingId,
-                httpFileService,
-                downloadHttpTimeoutInMilliseconds,
-                zipExtractor,
+                requestId, trackingId,
+                httpFileService, downloadHttpTimeoutInMilliseconds, zipExtractor,
                 new DownloadStatusProvider(requestId, serviceClient, statusPollIntervalInMilliseconds));
     }
